🔧 (example/.gitignore): update .gitignore to include additional directories and files The .gitignore file is updated to exclude more directories and files that are commonly generated by development tools and should not be tracked in version control. This includes directories like .expo, .vscode, and various build directories for Android and iOS, as well as configuration files like jsconfig.json. This change helps keep the repository clean and focused on source code by ignoring files that are environment-specific or generated during the build process.
⬆️ (dependencies): add new dependencies to yarn cache

Add a set of new dependencies to the yarn cache to ensure that the
project has all the necessary packages for its functionality. This
change is crucial for maintaining the integrity and performance of
the project by ensuring that all required packages are available
locally, reducing the need for repeated downloads and improving
build times.

🔧 (example): update project configuration and dependencies for improved build and compatibility

Update the Android namespace and application ID to follow a more standard naming convention. Add new architecture support in the Android build configuration to enable code generation when the new architecture is enabled. Update the package name in `package.json` to align with the new naming convention and increment the version for release tracking. Upgrade dependencies to their latest versions to ensure compatibility and access to new features. Replace custom Babel and Metro configurations with those provided by `react-native-builder-bob` for consistency and maintainability. Remove unused TypeScript configuration file to clean up the project structure. These changes enhance the project's maintainability, compatibility, and readiness for future updates.